DS-232-menu_update_fix
